Add brand icon, pin ruff rules, fix lint
The HACS validation needs repository topics and a brand icon as long as
the integration is not in the official brands repository. The icon is
generated, not drawn: a sun over a charging plug.

Ruff rules are now pinned in pyproject.toml. Without that list the rules
depend on whichever version happens to be installed, and the run fails
exactly when it is in the way - which is what happened on the first push.

+++ custom_components/pv_surplus/actuator.py
@@ -64,7 +64,7 @@ class Actuator:
     def reported_current(self) -> int | None:
         """Zurueckgemeldetes Limit der Stell-Entitaet."""
         wert = read_number(self._hass.states.get(self._current_entity))
-        return None if wert is None else int(round(wert))
+        return None if wert is None else round(wert)
 
     def external_change(self, now: float) -> bool:
         """Hat jemand anderes das Limit veraendert?
